Ask for confirmation before deleting Support for `-Force` switch Improve default value for Tenant parameter Enhanced verbose output

+++ scripts/PowerShell/Scripts/Remove-Spoke.ps1
@@ -11,6 +11,12 @@
 .PARAMETER CloudEnvironment
     The Azure environment where the spoke was created. Default is 'AzureCloud'.
 
+.PARAMETER Tenant
+    The Azure tenant ID where the spoke was created. Default is the current tenant.
+
+.PARAMETER Force
+    DANGER: Forces the deletion of the spoke resources without prompting for confirmation.
+
 .EXAMPLE
     ./deploy.ps1 -TemplateParameterFile '.\main.hub.bicepparam' -TargetSubscriptionId '00000000-0000-0000-0000-000000000000'
 
@@ -25,7 +31,7 @@
 #Requires -Modules Az.Resources, Az.RecoveryServices, Az.Network, Az.DataFactory
 #Requires -PSEdition Core
 
-[CmdletBinding(SupportsShouldProcess)]
+[CmdletBinding(SupportsShouldProcess = $true)]
 param (
     [Parameter(Mandatory, Position = 1)]
     [string]$TemplateParameterFile,
@@ -34,7 +40,9 @@ param (
     [Parameter(Position = 3)]
     [string]$CloudEnvironment = 'AzureCloud',
     [Parameter(Position = 4)]
-    [string]$TenantId = ''
+    [string]$Tenant = (Get-AzContext).Tenant.Id,
+    [Parameter()]
+    [switch]$Force
 )
 
 $ErrorActionPreference = 'Stop'
@@ -73,9 +81,8 @@ try {
     Import-Module ..\Modules\AzSubscriptionManagement.psm1
 
     $OriginalContext = Get-AzContext
-
     # Determine if a cloud context switch is required
-    $AzContext = Set-AzContextWrapper -SubscriptionId $TargetSubscriptionId -Environment $CloudEnvironment
+    $AzContext = Set-AzContextWrapper -SubscriptionId $TargetSubscriptionId -Environment $CloudEnvironment -Tenant $Tenant
 
     # Check if any resource groups exist that match the pattern
     $ResourceGroups = Get-AzResourceGroup -Name $ResourceGroupNamePattern
@@ -84,14 +91,28 @@ try {
         Write-Warning "No resource groups found matching pattern '$ResourceGroupNamePattern' in subscription '$((Get-AzContext).Subscription.Name)'."
         exit
     }
-    else {
-        Write-Host "Found $($ResourceGroups.Count) resource groups matching pattern '$ResourceGroupNamePattern' in subscription '$((Get-AzContext).Subscription.Name)'."
+
+    $Msg1 = "Found $($ResourceGroups.Count) resource groups matching pattern '$ResourceGroupNamePattern' in subscription '$((Get-AzContext).Subscription.Name)'."
+    $Msg = "$Msg1`nThese actions cannot be undone and data loss might occur. Do you want to continue?"
+
+    if (-not ($WhatIfPreference -or $Force -or $PSCmdlet.ShouldContinue($Msg, 'Confirm Spoke Removal'))) {    
+        exit
     }
+    
+    # If -WhatIf is used, output the number of resource groups found
+    if ($WhatIfPreference) {
+        Write-Host $Msg1
+    }
+
+    if ($Force) {
+        Write-Verbose "Force switch specified. Proceeding with deletion of resources."
+    }   
 
     ################################################################################
     # REMOVE ANY RESOURCE LOCKS
     ################################################################################
 
+    Write-Verbose "STEP 1: Removing resource locks..."
     $ResourceGroups | ForEach-Object { 
         Get-AzResourceLock -ResourceGroupName $_.ResourceGroupName | Remove-AzResourceLock -Force 
     }
@@ -107,7 +128,7 @@ try {
     $Vault = Get-AzRecoveryServicesVault -ResourceGroupName $BackupResourceGroupName -Name $RecoveryServicesVaultName -ErrorAction SilentlyContinue
 
     if ($Vault) {
-        Write-Verbose "Removing Recovery Services Vault '$RecoveryServicesVaultName' in resource group '$BackupResourceGroupName'."
+        Write-Verbose "STEP 2: Removing Recovery Services Vault '$RecoveryServicesVaultName' in resource group '$BackupResourceGroupName'..."
         & ./Recovery/Remove-rsv.ps1 -VaultName $RecoveryServicesVaultName `
             -ResourceGroup $BackupResourceGroupName -SubscriptionId $TargetSubscriptionId `
             -WhatIf:$WhatIfPreference -Verbose:$VerbosePreference
@@ -124,6 +145,7 @@ try {
     $Factory = Get-AzDataFactoryV2 -ResourceGroupName $StorageResourceGroupName -Name $DataFactoryName -ErrorAction SilentlyContinue
 
     if ($Factory) {
+        Write-Verbose "STEP 3: Removing managed private endpoints from Data Factory '$DataFactoryName' in resource group '$StorageResourceGroupName'..."
         & ./DataFactory/Remove-ManagedPrivateEndpoints.ps1 -DataFactoryName $DataFactoryName `
             -ResourceGroup $StorageResourceGroupName -SubscriptionId $TargetSubscriptionId `
             -WhatIf:$WhatIfPreference -Verbose:$VerbosePreference
@@ -136,6 +158,7 @@ try {
     # Two separate commands needed because -AsJob does not support specifying a variable
     if ($PSCmdlet.ShouldProcess("spoke resource groups", "Remove")) {
         $Jobs = @()
+        Write-Verbose "STEP 4: Removing resource groups..."
         $Jobs += $ResourceGroups | Remove-AzResourceGroup -AsJob -Force -Verbose:$VerbosePreference
 
         Write-Host "Waiting for $($Jobs.Count) resource groups to be deleted..."
@@ -162,9 +185,9 @@ try {
         [string]$NetworkResourceGroupName = $ResourceGroupNamePattern.Replace('*', 'network')
         [string]$SpokeVirtualNetworkResourceId = "/subscriptions/$TargetSubscriptionId/resourceGroups/$NetworkResourceGroupName/providers/Microsoft.Network/virtualNetworks/$SpokeVirtualNetworkName"
 
-        Write-Verbose "Removing disconnected peering to spoke network '$SpokeVirtualNetworkName' from hub virtual network '$HubVirtualNetworkName' in resource group '$HubResourceGroupName' in subscription '$HubSubscriptionId'."
+        Write-Verbose "STEP 5: Removing disconnected peering to spoke network '$SpokeVirtualNetworkName' from hub virtual network '$HubVirtualNetworkName' in resource group '$HubResourceGroupName' in subscription '$HubSubscriptionId'..."
     
-        $AzContext = Set-AzContextWrapper -SubscriptionId $HubSubscriptionId -Environment $CloudEnvironment    
+        $AzContext = Set-AzContextWrapper -SubscriptionId $HubSubscriptionId -Environment $CloudEnvironment -Tenant $Tenant
 
         # Remove peering explicitly from hub
         Get-AzVirtualNetworkPeering -ResourceGroupName $HubResourceGroupName -VirtualNetworkName $HubVirtualNetworkName | `
@@ -176,10 +199,13 @@ try {
 catch {
     Write-Host "An error occurred: $($_)"
     Write-Host $_.ScriptStackTrace
+    Write-Host "In context $AzContext"
 }
 finally {
-    $AzContext = Set-AzContextWrapper -SubscriptionId $OriginalContext.Subscription.Id -Environment $OriginalContext.Environment.Name
+    Write-Verbose "Setting Azure context back to the original subscription..."
+    $AzContext = Set-AzContextWrapper -SubscriptionId $OriginalContext.Subscription.Id -Environment $OriginalContext.Environment.Name -Tenant $OriginalContext.Tenant.Id
 
     # Remove the module from the session
     Remove-Module AzSubscriptionManagement -WhatIf:$false
+    Write-Verbose "Done!"
 }
